+++ include/arp.h
@@ -1,5 +1,5 @@
 #pragma once
-void arp_mode_changed();
+void reset_arp();
 void arpSteptrigger(int number);
 void arpReset(byte note);
 void arpTick();
