+++ include/version.h
@@ -1,0 +1,3 @@
+#pragma once
+static const int version = 3;
+static const int versionDecimal = 0;
