+++ include/voice_state.hpp
@@ -1,3 +1,4 @@
+#pragma once
 #include "voice_allocation.hpp"
 
 const long ONE_FP = 1L << 16;
